update: check if team member already exist on actions

+++ src/actions/deleteMemberAction.ts
@@ -1,28 +1,34 @@
 'use server'
 
 import { revalidateTag } from "next/cache"
+import getOneMemberAction from "./getOneMemberAction"
 
 const deleteMemberAction = async (id: number) => {
   try {
-    const res = await fetch(`http://localhost:3007/team-members/${id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    })
+    const member = await getOneMemberAction(id)
 
-    if(!res.ok) {
-      throw new Error("Failed to delete member!")
-    }
+    if (member) {
+      const res = await fetch(`http://localhost:3007/team-members/${id}`, {
+        cache: "no-cache",
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      })
 
-    revalidateTag("team_members")
-    const data = await res.json()
-    return {
-      success: "Success, team member deleted!",
-      data
+      if (!res.ok) {
+        throw new Error("Failed to delete member!")
+      }
+
+      revalidateTag("team_members")
+      const data = await res.json()
+      return {
+        success: "Success, team member deleted!",
+        data
+      }
     }
   }
-  catch(error: any) {
+  catch (error: any) {
     return {
       errors: error.toString()
     }
